Read GT3X

Read activity samples from a GT3X file as a matrix. Please note that all timestams are in local time (of the device) even though they are represented as POSIXct with GMT timezone.

read.gt3x(
  path,
  verbose = FALSE,
  asDataFrame = FALSE,
  imputeZeroes = FALSE,
  cleanup = FALSE,
  ...
)

Arguments

path
Path to gt3x folder

verbose
print diagnostic messages

asDataFrame
convert to an activity_df, see as.data.frame.activity

imputeZeroes
Impute zeros in case there are missingness? Default is FALSE, in which case the time series will be incomplete in case there is missingness.

cleanup
should any unzipped files be deleted?

...
additional arguments to pass to parseGT3X C++ code

Value

A numeric matrix with 3 columns (X, Y, Z) and the following attributes:

start_time : Start time from info file in POSIXct format.

subject_name : Subject name from info file

time_zone : Time zone from info file

missingness : Named integer vector. Names are Posixct timestamps and values are the number of missing values.

Note

The timestamps in the .gt3x data format are saved in .NET format, which is nanoseconds in local time since 0001-01-01. This is a bit tricky to parse into an R datetime format. DateTimes are therefore represented as POSIXct format with the 'GMT' timezone attribute, which is false; the datetime actually represents local time.

See also

Examples

# first unzip, then read
datadir <- gt3x_datapath()
gt3xfolders <- unzip.gt3x(datadir)

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf

#> 1/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt, log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/EE_left_29.5.2017-05-30

#> 2/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/.read.gt3x-data/SS_left_19.5.2017-05-22.gt3x

#>  === info.txt, log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/SS_left_19.5.2017-05-22

gt3xfile <- gt3xfolders[2]
is_gt3x(gt3xfile)

#> /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/SS_left_19.5.2017-05-22 
#>                                                                                FALSE 

have_log_and_info(gt3xfile, verbose = TRUE)

#> [1] TRUE

x <- read.gt3x(gt3xfile, imputeZeroes = TRUE, asDataFrame = FALSE,
verbose = TRUE)

#> GT3X information
#>  $ Serial Number     :"TAS1F28160235"
#>  $ Device Type       :"Link"
#>  $ Firmware          :"1.7.0"
#>  $ Battery Voltage   :"3,93"
#>  $ Sample Rate       :100
#>  $ Start Date        : POSIXct, format: "2017-05-19 10:50:00"
#>  $ Stop Date         : POSIXct, format: "0001-01-01"
#>  $ Last Sample Time  : POSIXct, format: "2017-05-22 15:12:34"
#>  $ TimeZone          :"03:00:00"
#>  $ Download Date     : POSIXct, format: "2017-05-22 15:12:34"
#>  $ Board Revision    :"6"
#>  $ Unexpected Resets :"0"
#>  $ Acceleration Scale:256
#>  $ Acceleration Min  :"-8.0"
#>  $ Acceleration Max  :"8.0"
#>  $ Limb              :"Wrist"
#>  $ Dominance         :"Non-Dominant"
#>  $ Subject Name      :"SS_left_19.5."
#>  $ Serial Prefix     :"TAS"

#> Parsing GT3X data via CPP.. expected sample size: 27495400

#> ---GT3X PARAMETERS
#> address: 0 key: 6 value: 0
#> address: 0 key: 7 value: 54686384
#> address: 0 key: 8 value: 6
#> address: 0 key: 9 value: 1469522671
#> address: 0 key: 13 value: 17235968
#> address: 0 key: 16 value: 3833593856
#> address: 0 key: 20 value: 0
#> address: 0 key: 21 value: 0
#> address: 0 key: 22 value: 0
#> address: 0 key: 23 value: 0
#> address: 0 key: 26 value: 2
#> address: 0 key: 28 value: 262013
#> address: 0 key: 29 value: 255
#> address: 0 key: 32 value: 16908288
#> address: 0 key: 37 value: 1024
#> address: 0 key: 38 value: 0
#> address: 0 key: 49 value: 2048
#> address: 0 key: 50 value: 88181047
#> address: 0 key: 51 value: 6.82667
#> address: 0 key: 55 value: 256
#> address: 0 key: 57 value: 333.87
#> address: 0 key: 58 value: 21
#> address: 0 key: 61 value: 2
#> address: 1 key: 0 value: 0
#> address: 1 key: 1 value: 3501572829
#> address: 1 key: 2 value: 384
#> address: 1 key: 3 value: 3
#> address: 1 key: 4 value: 4294967132
#> address: 1 key: 5 value: 4294967086
#> address: 1 key: 6 value: 4294967139
#> address: 1 key: 7 value: 286
#> address: 1 key: 8 value: 304
#> address: 1 key: 9 value: 326
#> address: 1 key: 10 value: 100
#> address: 1 key: 12 (start time)  value: 1495191000
#> address: 1 key: 13 value: 0
#> address: 1 key: 14 value: 1495190807
#> address: 1 key: 15 value: 65
#> address: 1 key: 16 value: 47
#> address: 1 key: 17 value: 81
#> address: 1 key: 20 value: 0
#> address: 1 key: 21 value: 0
#> address: 1 key: 33 value: 60000
#> address: 1 key: 34 value: 4294965227
#> address: 1 key: 35 value: 4294965203
#> address: 1 key: 36 value: 4294965245
#> address: 1 key: 37 value: 2030
#> address: 1 key: 38 value: 2003
#> address: 1 key: 39 value: 2048
#> address: 1 key: 40 value: 0
#> address: 1 key: 41 value: 0
#> address: 1 key: 42 value: 0
#> address: 1 key: 43 value: 4294967271
#> address: 1 key: 44 value: 0
#> address: 1 key: 45 value: 0
#> address: 1 key: 46 value: 0
#> ---END PARAMETERS
#> 
#> Total Records: 27489400
#> Scaling...
#> Creating dimnames 
#> CPP returning 

#> Done (in 3.50347018241882 seconds)

df2 <- as.data.frame(x)
head(df2)

#> Sampling Rate: 100Hz
#> Firmware Version: 1.7.0
#> Serial Number Prefix: TAS
#>       X     Y     Z                time
#> 1 0.961 0.027 0.160 2017-05-19 10:50:00
#> 2 0.980 0.027 0.172 2017-05-19 10:50:00
#> 3 0.988 0.031 0.168 2017-05-19 10:50:00
#> 4 0.992 0.031 0.168 2017-05-19 10:50:00
#> 5 0.992 0.031 0.172 2017-05-19 10:50:00
#> 6 0.992 0.027 0.176 2017-05-19 10:50:00

rm(x); gc(); gc()

#>             used  (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells    770708  41.2    1315955   70.3         NA   1315955   70.3
#> Vcells 111429432 850.2  358883047 2738.1      16384 358883047 2738.1

#>             used  (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells    770723  41.2    1315955   70.3         NA   1315955   70.3
#> Vcells 111429529 850.2  358883047 2738.1      16384 358883047 2738.1

rm(df2); gc()

#>           used (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells  770621 41.2    1315955   70.3         NA   1315955   70.3
#> Vcells 1447751 11.1  287106438 2190.5      16384 358883047 2738.1

# temporary unzip, read, convert to a data frame
gt3xfile <- gt3x_datapath(1)
memory.limit()

#> Warning: 'memory.limit()' is Windows-specific

#> [1] Inf

df <- read.gt3x(gt3xfile, asDataFrame = FALSE, verbose = 2)

#> Input is a .gt3x file, unzipping to a temporary location first...

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf

#> 1/1 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt, log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/EE_left_29.5.2017-05-30

#> GT3X information
#>  $ Serial Number     :"TAS1F27160060"
#>  $ Device Type       :"Link"
#>  $ Firmware          :"1.7.0"
#>  $ Battery Voltage   :"3,86"
#>  $ Sample Rate       :100
#>  $ Start Date        : POSIXct, format: "2017-05-23 09:15:00"
#>  $ Stop Date         : POSIXct, format: "0001-01-01"
#>  $ Last Sample Time  : POSIXct, format: "2017-05-30 09:50:29"
#>  $ TimeZone          :"03:00:00"
#>  $ Download Date     : POSIXct, format: "2017-05-30 09:50:29"
#>  $ Board Revision    :"6"
#>  $ Unexpected Resets :"0"
#>  $ Acceleration Scale:256
#>  $ Acceleration Min  :"-8.0"
#>  $ Acceleration Max  :"8.0"
#>  $ Limb              :"Wrist"
#>  $ Dominance         :"Non-Dominant"
#>  $ Subject Name      :"EE_left_29.5."
#>  $ Serial Prefix     :"TAS"

#> Parsing GT3X data via CPP.. expected sample size: 60692900

#> ---GT3X PARAMETERS
#> address: 0 key: 6 value: 0
#> address: 0 key: 7 value: 54727278
#> address: 0 key: 8 value: 6
#> address: 0 key: 9 value: 1469440996
#> address: 0 key: 13 value: 17235968
#> address: 0 key: 16 value: 3833593856
#> address: 0 key: 20 value: 0
#> address: 0 key: 21 value: 0
#> address: 0 key: 22 value: 0
#> address: 0 key: 23 value: 0
#> address: 0 key: 26 value: 2
#> address: 0 key: 28 value: 262013
#> address: 0 key: 29 value: 255
#> address: 0 key: 32 value: 16908288
#> address: 0 key: 37 value: 1027
#> address: 0 key: 38 value: 0
#> address: 0 key: 49 value: 2048
#> address: 0 key: 50 value: 88181047
#> address: 0 key: 51 value: 6.82667
#> address: 0 key: 55 value: 256
#> address: 0 key: 57 value: 333.87
#> address: 0 key: 58 value: 21
#> address: 0 key: 61 value: 2
#> address: 1 key: 0 value: 0
#> address: 1 key: 1 value: 3501572829
#> address: 1 key: 2 value: 384
#> address: 1 key: 3 value: 3
#> address: 1 key: 4 value: 4294967108
#> address: 1 key: 5 value: 4294967066
#> address: 1 key: 6 value: 4294967125
#> address: 1 key: 7 value: 302
#> address: 1 key: 8 value: 286
#> address: 1 key: 9 value: 314
#> address: 1 key: 10 value: 100
#> address: 1 key: 12 (start time)  value: 1495530900
#> address: 1 key: 13 value: 0
#> address: 1 key: 14 value: 1495530089
#> address: 1 key: 15 value: 62
#> address: 1 key: 16 value: 30
#> address: 1 key: 17 value: 66
#> address: 1 key: 20 value: 0
#> address: 1 key: 21 value: 0
#> address: 1 key: 33 value: 60000
#> address: 1 key: 34 value: 4294965246
#> address: 1 key: 35 value: 4294965214
#> address: 1 key: 36 value: 4294965227
#> address: 1 key: 37 value: 2056
#> address: 1 key: 38 value: 2020
#> address: 1 key: 39 value: 2048
#> address: 1 key: 40 value: 0
#> address: 1 key: 41 value: 4294967295
#> address: 1 key: 42 value: 0
#> address: 1 key: 43 value: 4294967252
#> address: 1 key: 44 value: 0
#> address: 1 key: 45 value: 0
#> address: 1 key: 46 value: 0
#> ---END PARAMETERS
#> 
#> Total Records: 60684200
#> Scaling...
#> Removing excess rows 
#> Sum of missingness is: 400
#> Finding missingness amount: 8300
#> Creating dimnames 
#> CPP returning 

#> Activity data now in R

#> Adding attributes

#> Done (in 12.4103388786316 seconds)

head(df)

#> Sampling Rate: 100Hz
#> Firmware Version: 1.7.0
#> Serial Number Prefix: TAS
#>           X     Y      Z
#> [1,] -0.035 0.348 -0.965
#> [2,] -0.008 0.352 -0.934
#> [3,]  0.008 0.332 -0.887
#> [4,]  0.008 0.320 -0.848
#> [5,]  0.004 0.316 -0.844
#> [6,]  0.000 0.301 -0.852

rm(df); gc(); gc()

#>           used (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells  771346 41.2    1315955   70.3         NA   1315955   70.3
#> Vcells 1449624 11.1  458181348 3495.7      16384 426270872 3252.2

#>           used (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells  771355 41.2    1315955   70.3         NA   1315955   70.3
#> Vcells 1449716 11.1  366545079 2796.6      16384 426270872 3252.2

df <- read.gt3x(gt3xfile, asDataFrame = TRUE, verbose = 2)

#> Input is a .gt3x file, unzipping to a temporary location first...

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf

#> 1/1 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt, log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/EE_left_29.5.2017-05-30

#> GT3X information
#>  $ Serial Number     :"TAS1F27160060"
#>  $ Device Type       :"Link"
#>  $ Firmware          :"1.7.0"
#>  $ Battery Voltage   :"3,86"
#>  $ Sample Rate       :100
#>  $ Start Date        : POSIXct, format: "2017-05-23 09:15:00"
#>  $ Stop Date         : POSIXct, format: "0001-01-01"
#>  $ Last Sample Time  : POSIXct, format: "2017-05-30 09:50:29"
#>  $ TimeZone          :"03:00:00"
#>  $ Download Date     : POSIXct, format: "2017-05-30 09:50:29"
#>  $ Board Revision    :"6"
#>  $ Unexpected Resets :"0"
#>  $ Acceleration Scale:256
#>  $ Acceleration Min  :"-8.0"
#>  $ Acceleration Max  :"8.0"
#>  $ Limb              :"Wrist"
#>  $ Dominance         :"Non-Dominant"
#>  $ Subject Name      :"EE_left_29.5."
#>  $ Serial Prefix     :"TAS"

#> Parsing GT3X data via CPP.. expected sample size: 60692900

#> ---GT3X PARAMETERS
#> address: 0 key: 6 value: 0
#> address: 0 key: 7 value: 54727278
#> address: 0 key: 8 value: 6
#> address: 0 key: 9 value: 1469440996
#> address: 0 key: 13 value: 17235968
#> address: 0 key: 16 value: 3833593856
#> address: 0 key: 20 value: 0
#> address: 0 key: 21 value: 0
#> address: 0 key: 22 value: 0
#> address: 0 key: 23 value: 0
#> address: 0 key: 26 value: 2
#> address: 0 key: 28 value: 262013
#> address: 0 key: 29 value: 255
#> address: 0 key: 32 value: 16908288
#> address: 0 key: 37 value: 1027
#> address: 0 key: 38 value: 0
#> address: 0 key: 49 value: 2048
#> address: 0 key: 50 value: 88181047
#> address: 0 key: 51 value: 6.82667
#> address: 0 key: 55 value: 256
#> address: 0 key: 57 value: 333.87
#> address: 0 key: 58 value: 21
#> address: 0 key: 61 value: 2
#> address: 1 key: 0 value: 0
#> address: 1 key: 1 value: 3501572829
#> address: 1 key: 2 value: 384
#> address: 1 key: 3 value: 3
#> address: 1 key: 4 value: 4294967108
#> address: 1 key: 5 value: 4294967066
#> address: 1 key: 6 value: 4294967125
#> address: 1 key: 7 value: 302
#> address: 1 key: 8 value: 286
#> address: 1 key: 9 value: 314
#> address: 1 key: 10 value: 100
#> address: 1 key: 12 (start time)  value: 1495530900
#> address: 1 key: 13 value: 0
#> address: 1 key: 14 value: 1495530089
#> address: 1 key: 15 value: 62
#> address: 1 key: 16 value: 30
#> address: 1 key: 17 value: 66
#> address: 1 key: 20 value: 0
#> address: 1 key: 21 value: 0
#> address: 1 key: 33 value: 60000
#> address: 1 key: 34 value: 4294965246
#> address: 1 key: 35 value: 4294965214
#> address: 1 key: 36 value: 4294965227
#> address: 1 key: 37 value: 2056
#> address: 1 key: 38 value: 2020
#> address: 1 key: 39 value: 2048
#> address: 1 key: 40 value: 0
#> address: 1 key: 41 value: 4294967295
#> address: 1 key: 42 value: 0
#> address: 1 key: 43 value: 4294967252
#> address: 1 key: 44 value: 0
#> address: 1 key: 45 value: 0
#> address: 1 key: 46 value: 0
#> ---END PARAMETERS
#> 
#> Total Records: 60684200
#> Scaling...
#> Removing excess rows 
#> Sum of missingness is: 400
#> Finding missingness amount: 8300
#> Creating dimnames 
#> CPP returning 

#> Activity data now in R

#> Adding attributes

#> Done (in 11.6047468185425 seconds)

#> Converting to a data.frame ...

#> Done

head(df)

#> Sampling Rate: 100Hz
#> Firmware Version: 1.7.0
#> Serial Number Prefix: TAS
#>        X     Y      Z                time
#> 1 -0.035 0.348 -0.965 2017-05-23 09:15:00
#> 2 -0.008 0.352 -0.934 2017-05-23 09:15:00
#> 3  0.008 0.332 -0.887 2017-05-23 09:15:00
#> 4  0.008 0.320 -0.848 2017-05-23 09:15:00
#> 5  0.004 0.316 -0.844 2017-05-23 09:15:00
#> 6  0.000 0.301 -0.852 2017-05-23 09:15:00

# \donttest{

url = "https://github.com/THLfi/read.gt3x/files/3522749/GT3X%2B.01.day.gt3x.zip"
destfile = tempfile(fileext = ".zip")
dl = download.file(url, destfile = destfile)
gt3x_file = unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file

res = read.gt3x(path)

gz = R.utils::gzip(path, remove = FALSE, overwrite = FALSE)
df2 <- read.gt3x(gz, asDataFrame = FALSE, verbose = 2)

#> Input is a .gt3x file, unzipping to a temporary location first...

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf

#> 1/1 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpbIvyOf/GT3X+ (01 day).gt3x

#>  === info.txt, activity.bin, lux.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpbIvyOf/GT3X+(01day)

#> GT3X information
#>  $ Serial Number     :"NEO1DXXXXXXXX"
#>  $ Device Type       :"GT3XPlus"
#>  $ Firmware          :"2.5.0"
#>  $ Battery Voltage   :"4.22"
#>  $ Sample Rate       :30
#>  $ Start Date        : POSIXct, format: "2012-06-27 10:54:00"
#>  $ Stop Date         : POSIXct, format: "2012-06-28 11:54:00"
#>  $ Download Date     : POSIXct, format: "2012-06-28 16:25:52"
#>  $ Board Revision    :"4"
#>  $ Unexpected Resets :"0"
#>  $ Sex               :"Male"
#>  $ Height            :"172.72"
#>  $ Mass              :"69.8532249799612"
#>  $ Age               :"43"
#>  $ Race              :"White / Caucasian"
#>  $ Limb              :"Ankle"
#>  $ Side              :"Left"
#>  $ Dominance         :"Non-Dominant"
#>  $ DateOfBirth       :"621132192000000000"
#>  $ Subject Name      :"GT3XPlus"
#>  $ Serial Prefix     :"NEO"
#>  $ Last Sample Time  : 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr "GMT"
#>  $ Acceleration Scale:341

#> Parsing GT3X data via CPP.. expected sample size: 2700000

#> Using NHANES-GT3X format - older format

#> Sample size: 2700000
#> Scaling...

#> Activity data now in R

#> Lux Sample size: 2700000

#> Adding attributes

#> Done (in 0.543074131011963 seconds)

head(df2)

#> Sampling Rate: 30Hz
#> Firmware Version: 2.5.0
#> Serial Number Prefix: NEO
#>      X Y Z
#> [1,] 0 0 0
#> [2,] 0 0 0
#> [3,] 0 0 0
#> [4,] 0 0 0
#> [5,] 0 0 0
#> [6,] 0 0 0

rm(df2); gc(); gc()

#>             used   (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells    816271   43.6    1315955   70.3         NA   1315955   70.3
#> Vcells 256409317 1956.3  790345861 6029.9      16384 790345861 6029.9

#>             used   (Mb) gc trigger   (Mb) limit (Mb)  max used   (Mb)
#> Ncells    815841   43.6    1315955   70.3         NA   1315955   70.3
#> Vcells 256408520 1956.3  790345861 6029.9      16384 790345861 6029.9

# }