Unzip gt3x files

unzip.gt3x() makes it convenient to unzip multiple .gt3x files.

unzip.gt3x(path, location = tempdir())

Arguments

path
One of the following: (1) A path to a directory with .gt3x files in which case they are all unzipped, or (2) A character vector of direct paths to .gt3x files.

location
Path to a directory to unzip the files to. Default is a temporary directory created by tempdir().

Value

Returns a vector of paths to unzipped gt3x folders.

Details

A .gt3x file is a zipped directory with two files: log.bin and info.txt. This function simply unzips the contents of the directories.

See also

Other file manipulations: gt3x_datapath, is_gt3x, list_gt3x

Examples

# NOT RUN {
# unzip a single .gt3x file
path <- gt3x_datapath(1)
gt3xdir <- unzip.gt3x(path)

# unzip multiple .gt3x files
dir <- gt3x_datapath()
gt3xdirs <- unzip.gt3x(dir)

# }