Unzip a single gt3x file

A .gt3x file is a zipped archive with two files: log.bin and info.txt. This function unzips the contents of the archive to a single folder. This is a helper for unzip.gt3x()

unzip_single_gt3x(path, dirname = basename(gsub(".gt3x$| ", "", path)),
  location = tempdir(), files = c("info.txt", "log.bin"))

Arguments

path
Path to a .gt3x file

dirname
The name of the resulting directory where the content of <path> are extracted. Default is the name of the input file, sans the .gt3x extension.

location
A path to an output directory. Default is a tempdir().

files
The names of files to extract. Default is info.txt and log.bin

See also

Other file manipulations: list_gt3x, unzip.gt3x