Read GT3X

Read activity samples from a GT3X file as a matrix. Please note that all timestams are in local time (of the device) even though they are represented as POSIXct with GMT timezone.

read.gt3x(path, verbose = FALSE, asDataFrame = FALSE,
  imputeZeroes = FALSE, ...)

Arguments

path
Path to gt3x folder

verbose
print diagnostic messages

asDataFrame
convert to an activity_df, see as.data.frame.activity

imputeZeroes
Impute zeros in case there are missingness? Default is FALSE, in which case the time series will be incomplete in case there is missingness.

...
additional arguments to pass to parseGT3X C++ code

Value

A numeric matrix with 3 columns (X, Y, Z) and the following attributes:

start_time : Start time from info file in POSIXct format.

subject_name : Subject name from info file

time_zone : Time zone from info file

missingness : Named integer vector. Names are Posixct timestamps and values are the number of missing values.

Note

The timestamps in the .gt3x data format are saved in .NET format, which is nanoseconds in local time since 0001-01-01. This is a bit tricky to parse into an R datetime format. DateTimes are therefore represented as POSIXct format with the 'GMT' timezone attribute, which is false; the datetime actually represents local time.

See also

Examples

# first unzip, then read
datadir <- gt3x_datapath()
gt3xfolders <- unzip.gt3x(datadir)

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu

#> 1/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/EE_left_29.5.2017-05-30

#> 2/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/SS_left_19.5.2017-05-22.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/SS_left_19.5.2017-05-22

gt3xfile <- gt3xfolders[2]
x <- read.gt3x(gt3xfile, imputeZeroes = TRUE)

#> Sample size: 27489200

df2 <- as.data.frame(x)
head(df2)

#>       X     Y     Z                time
#> 1 0.984 0.031 0.180 2017-05-19 10:50:01
#> 2 0.988 0.027 0.176 2017-05-19 10:50:01
#> 3 0.988 0.027 0.172 2017-05-19 10:50:01
#> 4 0.992 0.027 0.172 2017-05-19 10:50:01
#> 5 0.992 0.027 0.172 2017-05-19 10:50:01
#> 6 0.992 0.027 0.168 2017-05-19 10:50:01

# temporary unzip, read, convert to a data frame
gt3xfile <- gt3x_datapath(1)
df <- read.gt3x(gt3xfile, asDataFrame = TRUE)

#> Input is a .gt3x file, unzipping to a temporary location first...

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu

#> 1/1 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/EE_left_29.5.2017-05-30

#> Sample size: 60684100

head(df)

#>        X     Y      Z                time
#> 1 -0.164 0.820 -0.637 2017-05-23 09:15:01
#> 2 -0.145 0.766 -0.594 2017-05-23 09:15:01
#> 3 -0.121 0.766 -0.527 2017-05-23 09:15:01
#> 4 -0.094 0.777 -0.441 2017-05-23 09:15:01
#> 5 -0.090 0.777 -0.422 2017-05-23 09:15:01
#> 6 -0.098 0.754 -0.465 2017-05-23 09:15:01