Unzip gt3x files

unzip.gt3x() makes it convenient to unzip multiple .gt3x files.

unzip.gt3x(path, location = tempdir(), remove_original = FALSE)

Arguments

path
One of the following: (1) A path to a directory with .gt3x files in which case they are all unzipped, or (2) A character vector of direct paths to .gt3x files.

location
Path to a directory to unzip the files to. Default is a temporary directory created by tempdir().

remove_original
Remove the zipfiles after unzipping?

Value

Returns a vector of paths to unzipped gt3x folders.

Details

A .gt3x file is a zipped directory with two files: log.bin and info.txt. This function simply unzips the contents of the directories.

See also

Examples

# unzip a single .gt3x file
path <- gt3x_datapath(1)
gt3xdir <- unzip.gt3x(path)

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu

#> 1/1 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/EE_left_29.5.2017-05-30

# unzip multiple .gt3x files
dir <- gt3x_datapath()
gt3xdirs <- unzip.gt3x(dir)

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu

#> 1/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/EE_left_29.5.2017-05-30.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/EE_left_29.5.2017-05-30

#> 2/2 

#> Unzipping /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/.read.gt3x-data/SS_left_19.5.2017-05-22.gt3x

#>  === info.txt and log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpayLZPu/SS_left_19.5.2017-05-22