Unzip gt3x files

unzip.gt3x() makes it convenient to unzip multiple .gt3x files.

unzip.gt3x(path, verbose = TRUE, ...)

Arguments

path
One of the following: (1) A path to a directory with .gt3x files in which case they are all unzipped, or (2) A character vector of direct paths to .gt3x files.

verbose
print diagnostic messages

...
arguments to pass to unzip_single_gt3x

Value

Returns a vector of paths to unzipped gt3x folders.

Details

A .gt3x file is a zipped directory with two files: log.bin and info.txt. This function simply unzips the contents of the directories.

See also

Examples

gt3xfile <-
  system.file(
    "extdata", "TAS1H30182785_2019-09-17.gt3x",
    package = "read.gt3x")
gt3xdirs <- unzip.gt3x(gt3xfile)

#> Unzipping gt3x data to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpF9yy08

#> 1/1 

#> Unzipping /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpEaoXnd/temp_libpathd9ee53236a80/read.gt3x/extdata/TAS1H30182785_2019-09-17.gt3x

#>  === info.txt, log.bin extracted to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpF9yy08/TAS1H30182785_2019-09-17

if (FALSE) {
# unzip a single .gt3x file
path <- gt3x_datapath(1)
gt3xdir <- unzip.gt3x(path)

# unzip multiple .gt3x files
dir <- gt3x_datapath()
gt3xdirs <- unzip.gt3x(dir)
}
tfile = tempfile()
testthat::expect_error(unzip.gt3x(c(dir, tfile)))
testthat::expect_error(unzip.gt3x(c("", "")))